+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 ### Fixed
 
+* Don't use libavutil functions when configuring swresample context (fix #106)
 
 ## [1.3.5] - 2019-11-26
 ### Added